refactor(data): extract infused types into separate module
