+++ src/data/src/db/mod.rs
@@ -29,6 +29,9 @@ pub enum FoundOrInserted<T: Table> {
 /// Repository module
 pub mod repo;
 
+/// Entities with extra information embedded.
+pub mod infused;
+
 impl<T: Table> From<FoundOrInserted<T>> for Ref<T> {
     fn from(value: FoundOrInserted<T>) -> Self {
         match value {
